# server-go/docs/swagger.yaml
basePath: /api
definitions:
...
  models.AISessionResponse:
    description: Response containing AI session state
    properties:
      chart:
        allOf:
        - $ref: '#/definitions/models.Chart'
        description: Current chart state (draft)
      session:
        $ref: '#/definitions/models.AISession'
    type: object
...
  models.APIKey:
    properties:
      created:
        type: string
      expires_at:
        type: string
      id:
        description: UUID
        type: string
      last_used:
        type: string
      name:
        description: Human label
        type: string
      prefix:
        description: First 8 chars of the random token, plaintext
        type: string
      revoked:
        type: boolean
      revoked_at:
        type: string
      user_guid:
        description: Owner
        type: string
    type: object
...
  models.Chart:
    description: Chart/Control with data source binding, query config, and visualization
      settings
    properties:
      ai_session_id:
        description: Active AI session (drafts only)
        type: string
      chart_type:
        description: bar, line, pie, gauge, etc. (charts only)
        type: string
      component_code:
        description: React component code
        type: string
      component_type:
        description: '"chart" (default) | "control"'
        type: string
      connection_id:
        description: Reference to connection
        type: string
      control_config:
        allOf:
        - $ref: '#/definitions/models.ControlConfig'
        description: Control configuration (controls only)
      created:
        type: string
      data_mapping:
        allOf:
        - $ref: '#/definitions/models.ChartDataMapping'
        description: How to map data to chart (charts only)
      description:
        type: string
      display_config:
        allOf:
        - $ref: '#/definitions/models.DisplayConfig'
        description: Display configuration (displays only)
      id:
        description: UUID - same across versions
        type: string
      name:
        description: Unique identifier within namespace
        type: string
      namespace:
        description: Conflict-domain; uniqueness is (namespace, name). See models.Namespace.
        type: string
      options:
        additionalProperties: true
        description: ECharts options overrides (charts only)
        type: object
      query_config:
        allOf:
        - $ref: '#/definitions/models.ChartQueryConfig'
        description: How to query data (charts only)
      status:
        description: '"draft" | "final"'
        type: string
      tags:
        description: Searchable tags
        items:
          type: string
        type: array
      title:
        description: Display title (defaults to Name if empty)
        type: string
      updated:
        type: string
      use_custom_code:
        description: Whether custom code mode is enabled
        type: boolean
      version:
        description: Version number (1, 2, 3...)
        type: integer
    required:
    - name
    type: object
...
  models.ChartListResponse:
    description: Response containing a list of charts with pagination
    properties:
      charts:
        items:
          $ref: '#/definitions/models.Chart'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  models.ChartQueryConfig:
    description: Query configuration for fetching chart data
    properties:
      params:
        additionalProperties: true
        description: Query parameters
        type: object
      raw:
        description: SQL query, filter, or API path
        type: string
      type:
        description: sql, csv_filter, stream_filter, api
        type: string
    type: object
...
  models.CreateAISessionRequest:
    description: Request body for creating a new AI session
    properties:
      chart_id:
        description: Existing chart ID to edit (optional, omit for new chart)
        type: string
      chart_type:
        description: 'For charts: bar, line, pie, etc.'
        type: string
      component_type:
        description: Pre-flight context (optional) - sets fields on the draft directly
        type: string
      connection_id:
        description: Pre-selected connection ID
        type: string
      control_type:
        description: 'For controls: button, toggle, slider, etc.'
        type: string
      dashboard_id:
        description: Dashboard panel context (optional) - if set, saved component
          is auto-attached to this panel
        type: string
      initial_message:
        description: First user message (optional)
        type: string
      panel_id:
        description: Panel within the dashboard
        type: string
    type: object
...
  models.CreateChartRequest:
    description: Request body for creating a new chart or control
    properties:
      chart_type:
        type: string
      component_code:
        type: string
      component_type:
        description: '"chart" (default) | "control"'
        type: string
      connection_id:
        type: string
      control_config:
        $ref: '#/definitions/models.ControlConfig'
      data_mapping:
        $ref: '#/definitions/models.ChartDataMapping'
      description:
        type: string
      display_config:
        $ref: '#/definitions/models.DisplayConfig'
      name:
        type: string
      namespace:
        description: Empty defaults to "default" in the handler.
        type: string
      options:
        additionalProperties: true
        type: object
      query_config:
        $ref: '#/definitions/models.ChartQueryConfig'
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      use_custom_code:
        type: boolean
    required:
    - name
    type: object
  models.CreateDashboardRequest:
    description: Request body for creating a new dashboard
    properties:
      description:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      namespace:
        description: Empty defaults to "default" in the handler.
        type: string
      panels:
        description: Panels with optional chart_id
        items:
          $ref: '#/definitions/models.DashboardPanel'
        type: array
      settings:
        $ref: '#/definitions/models.DashboardSettings'
      tags:
        items:
          type: string
        type: array
    required:
    - name
    type: object
  models.CreateDatasourceRequest:
    properties:
      config:
        $ref: '#/definitions/models.DatasourceConfig'
      description:
        type: string
      mask_secrets:
        description: 'If true, secrets are masked in API responses (default: true)'
        type: boolean
      name:
        type: string
      namespace:
        description: Empty defaults to "default" in the handler.
        type: string
      supported_schemas:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      type:
        allOf:
        - $ref: '#/definitions/models.DatasourceType'
        description: 'LEGACY: Keep for backwards compatibility'
      type_config:
        additionalProperties: true
        type: object
      type_id:
        description: 'NEW: Registry-based type system (preferred)'
        type: string
    required:
    - name
    type: object
...
  models.CreateNamespaceRequest:
    description: Request body for creating a new namespace
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  models.CreateUserRequest:
    description: Request body for creating a new user
    properties:
      capabilities:
        items:
          $ref: '#/definitions/models.Capability'
        type: array
      email:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  models.Dashboard:
    description: Dashboard with panels that reference standalone charts
    properties:
      created:
        type: string
      description:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      namespace:
        description: Conflict-domain; uniqueness is (namespace, name). See models.Namespace.
        type: string
      panels:
        description: Panels with chart_id references
        items:
          $ref: '#/definitions/models.DashboardPanel'
        type: array
      settings:
        $ref: '#/definitions/models.DashboardSettings'
      tags:
        description: User-defined tags for filtering/grouping
        items:
          type: string
        type: array
      thumbnail:
        description: Base64 encoded thumbnail image
        type: string
      updated:
        type: string
    required:
    - name
    type: object
...
  models.DashboardSummaryListResponse:
    description: Response containing dashboard summaries with optional data source
      info
    properties:
      dashboards:
        items:
          $ref: '#/definitions/models.DashboardSummary'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  models.Datasource:
    properties:
      config:
        $ref: '#/definitions/models.DatasourceConfig'
      created_at:
        type: string
      description:
        type: string
      health:
        $ref: '#/definitions/models.HealthInfo'
      id:
        type: string
      mask_secrets:
        description: If true, secrets are masked in API responses
        type: boolean
      name:
        type: string
      namespace:
        description: Conflict-domain; uniqueness is (namespace, name). See models.Namespace.
        type: string
      supported_schemas:
        description: Control schema support - which control schemas this connection
          supports
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      type:
        allOf:
        - $ref: '#/definitions/models.DatasourceType'
        description: 'LEGACY: Keep for backwards compatibility during migration'
      type_config:
        additionalProperties: true
        type: object
      type_id:
        description: |-
          NEW: Registry-based type system (preferred)
          TypeID format: "category.name" (e.g., "db.postgres", "stream.websocket-bidir")
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
...
  models.Namespace:
    description: Namespace groups connections, components, and dashboards into conflict
      domains
    properties:
      color:
        description: Carbon-safe hex; UI picks from a fixed palette
        type: string
      created:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated:
        type: string
    type: object
  models.NamespaceListResponse:
    description: Response containing a list of namespaces
    properties:
      namespaces:
        items:
          $ref: '#/definitions/models.Namespace'
        type: array
      total:
        type: integer
    type: object
...
  models.SystemConfigResponse:
    properties:
      default_dimension:
        type: string
      layout_dimensions:
        additionalProperties:
          $ref: '#/definitions/models.LayoutDimensionDTO'
        type: object
      settings:
        additionalProperties: true
        type: object
    type: object
...
  models.UpdateConfigRequest:
    properties:
      settings:
        additionalProperties: true
        type: object
    required:
    - settings
    type: object
...
  models.User:
    description: User account with authentication and capabilities
    properties:
      active:
        description: Whether user is active
        type: boolean
      capabilities:
        description: User capabilities
        items:
          $ref: '#/definitions/models.Capability'
        type: array
      created:
        type: string
      email:
        description: Optional email
        type: string
      guid:
        description: UUID for authentication header
        type: string
      id:
        type: string
      name:
        description: Display name
        type: string
      updated:
        type: string
    type: object
...
  models.UserListResponse:
    description: Response containing a list of users with pagination
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
host: localhost:3001
info:
  contact:
    email: support@example.com
    name: Dashboard Team
  description: Dashboard system with AI-powered chart generation
  title: TRVE Dashboards API
  version: "1.0"
paths:
  /ai/sessions:
    post:
      consumes:
      - application/json
      description: Create a new AI session for chart creation or editing. Creates
        a chart draft.
      parameters:
      - description: Session creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAISessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AISessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new AI session
      tags:
      - ai
...
  /api/api-keys/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.APIKey'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List every API key in the deployment (admin)
      tags:
      - api-keys
...
  /api/namespaces:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NamespaceListResponse'
      summary: List namespaces
      tags:
      - namespaces
    post:
      consumes:
      - application/json
      parameters:
      - description: Namespace to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateNamespaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Namespace'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a namespace
      tags:
      - namespaces
...
  /charts:
    get:
      description: Get a paginated list of charts (latest version of each) with optional
        filtering
      parameters:
      - description: Filter by name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by chart type
        in: query
        name: chart_type
        type: string
      - description: Filter by component type (chart, control, display)
        in: query
        name: component_type
        type: string
      - description: Filter by status (draft, final)
        in: query
        name: status
        type: string
      - description: Filter by data source ID
        in: query
        name: datasource_id
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChartListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List charts
      tags:
      - charts
    post:
      consumes:
      - application/json
      description: Create a new chart with data source binding and visualization config.
        Creates as version 1 with status "final".
      parameters:
      - description: Chart data
        in: body
        name: chart
        required: true
        schema:
          $ref: '#/definitions/models.CreateChartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Chart'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new chart
      tags:
      - charts
...
  /config/system:
    get:
      description: Retrieves system-wide configuration including layout dimensions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SystemConfigResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get system configuration
      tags:
      - config
    put:
      consumes:
      - application/json
      description: Updates system-wide configuration settings
      parameters:
      - description: Configuration settings to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SystemConfigResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update system configuration
      tags:
      - config
...
  /dashboards:
    get:
      description: Get a paginated list of dashboards with optional filtering. Use
        include_datasources=true to get data source names for each dashboard.
      parameters:
      - description: Filter by name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by public status
        in: query
        name: is_public
        type: boolean
      - description: Include data source names from charts
        in: query
        name: include_datasources
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Response when include_datasources=true
          schema:
            $ref: '#/definitions/models.DashboardSummaryListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List dashboards
      tags:
      - dashboards
    post:
      consumes:
      - application/json
      description: Create a new dashboard with panels and embedded charts
      parameters:
      - description: Dashboard data
        in: body
        name: dashboard
        required: true
        schema:
          $ref: '#/definitions/models.CreateDashboardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Dashboard'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new dashboard
      tags:
      - dashboards
...
  /datasources:
    get:
      description: Retrieve all datasources with pagination and optional namespace/type/tag
        filters
      parameters:
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      - description: Filter by namespace (empty = all namespaces)
        in: query
        name: namespace
        type: string
      - description: Filter by datasource type (api, websocket, file)
        in: query
        name: type
        type: string
      - collectionFormat: csv
        description: Filter by tags (OR semantics, repeat param)
        in: query
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List all datasources
      tags:
      - datasources
    post:
      consumes:
      - application/json
      description: Create a new data source (API, WebSocket, or File)
      parameters:
      - description: Datasource to create
        in: body
        name: datasource
        required: true
        schema:
          $ref: '#/definitions/models.CreateDatasourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Datasource'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new datasource
      tags:
      - datasources
...
  /users:
    get:
      description: Returns a paginated list of all users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserListResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - Users
...
swagger: "2.0"